requirements.txt
BeautifulSoup==3.2.1
flake8==3.5.0
flake8-isort==2.3
mozlog==3.7
PyPOM==1.2.0
pytest==3.4.0
pytest-metadata==1.5.1
pytest-selenium==1.11.4
pytest-variables==1.7.1
pytest-xdist==1.22.0
requests==2.18.4
pyotp==2.2.6
